Use cask tab to better identify installed casks
Previously we would identify installed casks by token derived from
directories present in the caskroom directory. That was error prone
when the token was ambiguous. In that case, the `Cask::Caskroom.cask`
method would default to the first cask which was not always correct.

Now that we have the cask tab we can use the installed receipt to
get the tap a cask came from. With both the tap and the token, we
can be sure that we're loading the cask that is actually installed
and not another cask from a different tap that just happens to have
the same name.